+++ chap1/ft_memcmp.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned const char	*data1;
 	unsigned const char	*data2;
