+++ chap1/ft_memmove.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-void	*ft_memmove(void *dest, const void *src, unsigned int n)
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	unsigned int		i;
+	size_t			i;
 	unsigned const char	*data_in;
 	unsigned char	*data_to;
