+++ chap1/ft_strnstr.c
@@ -1,9 +1,9 @@
 #include "libft.h"
 
-char	*ft_strnstr(const char *big, const char *little, unsigned int len)
+char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t		i;
+	size_t		j;
 	char		*data_b;
 	char		*data_l;
